Make updater env restoration panic-safe in tests
