+++ updater/src/test_util.rs
@@ -18,3 +18,52 @@ pub(crate) fn env_lock() -> MutexGuard<'static, ()> {
         .lock()
         .unwrap_or_else(|err| err.into_inner())
 }
+
+pub(crate) struct EnvRestoreGuard {
+    saved: Vec<(&'static str, Option<std::ffi::OsString>)>,
+}
+
+impl EnvRestoreGuard {
+    pub(crate) fn capture(keys: &[&'static str]) -> Self {
+        Self {
+            saved: keys
+                .iter()
+                .map(|key| (*key, std::env::var_os(key)))
+                .collect(),
+        }
+    }
+}
+
+impl Drop for EnvRestoreGuard {
+    fn drop(&mut self) {
+        for (key, value) in &self.saved {
+            if let Some(value) = value {
+                std::env::set_var(key, value);
+            } else {
+                std::env::remove_var(key);
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{env_lock, EnvRestoreGuard};
+
+    #[test]
+    fn env_restore_guard_restores_variables_after_panic() {
+        let _env_guard = env_lock();
+        let original_display = std::env::var_os("DISPLAY");
+        let original_wayland_display = std::env::var_os("WAYLAND_DISPLAY");
+
+        let _ = std::panic::catch_unwind(|| {
+            let _restore_env = EnvRestoreGuard::capture(&["DISPLAY", "WAYLAND_DISPLAY"]);
+            std::env::set_var("DISPLAY", ":panic-test");
+            std::env::remove_var("WAYLAND_DISPLAY");
+            panic!("intentional panic to verify env restoration");
+        });
+
+        assert_eq!(std::env::var_os("DISPLAY"), original_display);
+        assert_eq!(std::env::var_os("WAYLAND_DISPLAY"), original_wayland_display);
+    }
+}
